requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
SQLAlchemy>=2.0
pydantic-settings>=2.0
email-validator>=2.0
python-jose[cryptography]>=3.3
passlib>=1.7.4
python-multipart>=0.0.9
pypdf>=4.0
pytest>=8.0
httpx>=0.27
psycopg[binary]
openai>=1.51.0
python-jose[cryptography]>=3.3.0